*visualisation fixes, add some tests, number work fix* major fixes: 	- fix allign in catalog: make single container for images and work placeholder minor fixes: 	- fix work with prices and dec values 	- fix tests for encryption, add new dependences

+++ app/features/products/schemas.py
@@ -8,7 +8,7 @@ class ProductC(BaseModel):
     name: str
     dimensions: Optional[str] = None
     unit: str
-    price: int
+    price: Decimal
     quantity_available: int = 0
     image_path: Optional[str] = None
